+++ egui/src/widget_text.rs
@@ -563,14 +563,14 @@ impl WidgetText {
             Self::RichText(text) => {
                 let valign = ui.layout().vertical_align();
                 let mut text_job = text.into_text_job(ui.style(), fallback_font.into(), valign);
-                text_job.job.wrap_width = wrap_width;
+                text_job.job.wrap.max_width = wrap_width;
                 WidgetTextGalley {
                     galley: ui.fonts().layout_job(text_job.job),
                     galley_has_color: text_job.job_has_color,
                 }
             }
             Self::LayoutJob(mut job) => {
-                job.wrap_width = wrap_width;
+                job.wrap.max_width = wrap_width;
                 WidgetTextGalley {
                     galley: ui.fonts().layout_job(job),
                     galley_has_color: true,