+++ egui_demo_lib/src/apps/demo/code_editor.rs
@@ -78,7 +78,7 @@ impl super::View for CodeEditor {
         let mut layouter = |ui: &egui::Ui, string: &str, wrap_width: f32| {
             let mut layout_job =
                 crate::syntax_highlighting::highlight(ui.ctx(), &theme, string, language);
-            layout_job.wrap_width = wrap_width;
+            layout_job.wrap.max_width = wrap_width;
             ui.fonts().layout_job(layout_job)
         };